I created four html files to make up the structure of the website. One index page and three recipe pages.
The index file contains a title and an unordered list contained within a "div" element. With 3 internal links directing the user to other recipes. Each recipe page contains
an unordered list of Ingredients, a ordered list of directions, a table with for rows, and a footer with links home and to other recipes