Fixed a link in beef-stew file, and added CSS to style the pages.
Styled the navbar on the index.html file, centered and spaced the links using flex. Also added background color to the html.
I centered the h1 and h2 headings and  added background color.
Created a bottom border to the h2 headings. Also centered the images and
rounded off the corners with a radius of 10px.